Raise `ArgumentError` if `Version` is initialized with a `nil` version string
At the moment, `Semantic::Version.new` raises a `NoMethodError`
exception like this if you try to pass in `nil` as the version
string:

```
undefined method `match' for nil:NilClass (NoMethodError)

v = version_str.match(SemVerRegexp)
```

This alters the behaviour to raise an `ArgumentError` instead,
which is what we do if you pass a malformed string (e.g.
`dasd`). This is a breaking change - but feels like a more
reasonable and unsurprising behaviour.